Adds an action for sharing photos elsewhere with a public cache
This is targetted towards high traffic from places like Reddit.
By using a short URL, e.g. /p/104, or appending /share to the photo's
standard URL, we can use a different action which can be cached for everyone.

An improvement to this will be to asyncronously load the user's correct
nav if they're signed in. At the moment it's cached as a signed-out user.